+++ common/buffer/ring.h
@@ -71,7 +71,7 @@ public: /*< ## Public Memeber Methods */
     inline u64    size()     { return m_bd->size; }
     inline u64    capacity() { return m_capacity; }
     inline u64    count()    { return m_capacity; }
-    inline c_cstr name()     { return m_bd->name; } 
+    inline c_cstr name()     { return m_bd->name; }
 
     inline void drop() {
         memset(m_bd->data, '\0', m_bd->size);
@@ -81,7 +81,7 @@ public: /*< ## Public Memeber Methods */
     inline u64 resize(u64 capacity) {
         N2CRASH_IF(m_resize == nullptr, N2Error::MissingData,
                    "Unable to resize ring %s (resize function not set)",
-                   name()); 
+                   name());
         auto required_size = Ring<T>::precomputeSize(capacity);
         auto new_size = m_resize(m_bd, required_size);
         m_capacity = new_size / sizeof(T);
@@ -171,7 +171,7 @@ protected: /*< ## Protected Member Methods */
 
     inline u64 decrement(u64 index, i64 n = 1) {
         if (n >= 0) {
-            while (index < n) {
+            while (index < (u64)n) {
                 index += capacity();
             }
             return index - n;
